index: add trace2 region for clear skip worktree
In a large repository using sparse checkout, checking whether a file marked
with skip worktree is present on disk and its skip worktree bit should be
cleared can take a considerable amount of time. Add a trace2 region to
surface this information, keeping a count of how many paths have been
checked and separately keep counts for after a full index is materialised.

Signed-off-by: Anh Le <anh@canva.com>