merge-index: migrate to parse_options() API
Migrate the "merge-index" command to the parse_options() API, a
preceding commit added tests for the existing behavior.

In a subsequent commit we'll adjust the behavior to be more consistent
with how most other commands work, but for now let's take pains to
preserve it as-is. We need to e.g. call parse_options() twice now, as
the "-a" option is currently only understood after "<merge-program>".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>