builtin/merge-index.c: don't USE_THE_INDEX_COMPATIBILITY_MACROS
Remove "USE_THE_INDEX_COMPATIBILITY_MACROS" and instead pass
"the_index" around between the functions in this file. In a subsequent
commit we'll libify this, and don't want to use
"USE_THE_INDEX_COMPATIBILITY_MACROS" in any more places in the
top-level *.c files. Doing this first makes that diff a lot smaller.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>