[RPC] grep: introduce --scope option
Because we may want to restrict grep's file scope
to sparse specification, Apply the --scope option
we implemented in diff to grep as well.

`--scope=sparse` mean that the search file scope
restrict to sparse specification when we grep
something in commit history, and `--scope=all`
mean that the search file scope will be full-tree.

Note that `--scope` option only oly takes effect
when "git grep <tree>" is specified.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>