mingw: remove msysGit/MSYS1 support
The msysGit (i.e. Git for Windows 1.x' SDK) is no longer used and the
last known user for MSys1 + MinGW has switched to MSys2. So there was no
reason to keep the msysGit/MSys1 config section. This Patch intends to
remove the support of both the builds and throw an error on
execution.

The output of `uname -r` if returns us with '1.' then it relates with MSys1
build environment and msysGit which we don't want to continue ahead and thus
an error can be expected.

The MSYS2 is currently used into the MinGW so we can get that part
running directly without any conditionals required.

Signed-off-by: Harshil-Jani <harshiljani2002@gmail.com>