doc: use "git switch -c" rather than "git checkout -b" consistently
In the "DETACHED HEAD" section in the git-checkout doc, it suggests
using "git checkout -b <branch-name>" to create a new branch on the
detached head.

On the other hand, when you checkout a commit that is not at the tip of
any named branch (e.g., when you checkout a tag), git suggests using
"git switch -c <branch-name>".

Use "git switch -c" and fix this inconsistency.

Signed-off-by: Yutaro Ohno <yutaro.ono.418@gmail.com>