config.c: plumb config_source through static fns
This reduces the direct dependence on the "cf" static variable, which
will make it easier to remove in a later commit. The plumbed arg is
named "cs" to differentiate between it and the static variable.

In some cases (public functions and config callback functions), there
isn't an obvious way to plumb "struct config_source" through function
args. As a workaround, add references to "cf" that we'll address in
later commits.

The remaining references to "cf" are direct assignments to "cf", which
we'll also address in a later commit.

Signed-off-by: Glen Choo <chooglen@google.com>